+++ src/Acl.Fs.Constant/Cryptography/CryptoConstants.cs
@@ -14,11 +14,6 @@ internal static class CryptoConstants
     internal static readonly byte[] NonceContext =
         [0x41, 0x43, 0x4C, 0x5F, 0x4E, 0x4F, 0x4E, 0x43, 0x45];
 
-    internal static int HmacKeySize => IsRunningOnGitHubActions ? 32 : 64;
-
-    internal static int SaltSize => IsRunningOnGitHubActions ? 32 : 64;
-
-    internal static bool IsRunningOnGitHubActions =>
-        Environment.GetEnvironmentVariable("GITHUB_ACTIONS") is "true" ||
-        Environment.GetEnvironmentVariable("CI") is "true";
+    internal static int HmacKeySize => 32;
+    internal static int SaltSize => 32;
 }