+++ src/Acl.Fs.Constant/Versioning/VersionConstants.cs
@@ -6,7 +6,7 @@ namespace Acl.Fs.Constant.Versioning;
 internal static class VersionConstants
 {
     internal const byte CurrentMajorVersion = 0;
-    internal const byte CurrentMinorVersion = 2; 
+    internal const byte CurrentMinorVersion = 3;
 
     internal const int VersionHeaderSize = 2;
 
@@ -14,6 +14,14 @@ internal static class VersionConstants
         VersionHeaderSize + CryptoConstants.NonceSize + sizeof(long) + CryptoConstants.SaltSize +
         CryptoConstants.Argon2IdSaltSize;
 
+    internal static int XChaCha20Poly1305UnalignedHeaderSize =>
+        VersionHeaderSize + CryptoConstants.XChaCha20Poly1305NonceSize + sizeof(long) + CryptoConstants.SaltSize +
+        CryptoConstants.Argon2IdSaltSize;
+
     internal static int HeaderSize => (UnalignedHeaderSize + StorageConstants.SectorSize - 1) /
+        StorageConstants.SectorSize * StorageConstants.SectorSize;
+
+    internal static int XChaCha20Poly1305HeaderSize =>
+        (XChaCha20Poly1305UnalignedHeaderSize + StorageConstants.SectorSize - 1) /
         StorageConstants.SectorSize * StorageConstants.SectorSize;
 }