+++ src/Acl.Fs.Core/Abstractions/Service/Decryption/Shared/Buffer/IBufferManager.cs
@@ -9,4 +9,5 @@ public interface IBufferManager : IDisposable
     byte[] Tag { get; }
     byte[] ChunkNonce { get; }
     byte[] Salt { get; }
+    int NonceSize { get; }
 }