+++ src/Acl.Fs.Core/Abstractions/Service/Decryption/Shared/Header/IHeaderReader.cs
@@ -8,4 +8,12 @@ internal interface IHeaderReader
         byte[] chaCha20Salt,
         int metadataBufferSize,
         CancellationToken cancellationToken);
+
+    Task<Header> ReadHeaderAsync(
+        System.IO.Stream sourceStream,
+        byte[] metadataBuffer,
+        byte[] chaCha20Salt,
+        int metadataBufferSize,
+        int nonceSize,
+        CancellationToken cancellationToken);
 }